+++ src/about.ts
@@ -2,24 +2,15 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import * as vscode from 'vscode';
-import * as path from 'node:path';
-import * as fs from 'node:fs';
 import * as cedar from 'vscode-cedar-wasm';
 
-const packageJsonFile = path.join(__dirname, '..', 'package.json');
-const packageJson = JSON.parse(fs.readFileSync(packageJsonFile, 'utf8'));
-const packageVersion = `${packageJson.publisher}.${packageJson.name}: ${packageJson.version}\n`;
-const vsCodeVersion = `Visual Studio Code: ${vscode.version}\n`;
-const nodeVersion = process.versions.node
-  ? `node: ${process.versions.node}\n`
-  : '';
-
-export async function aboutExtension(): Promise<void> {
+export async function aboutExtension(context: vscode.ExtensionContext): Promise<void> {
+  const pkg = context.extension.packageJSON;
   const extensionDetails =
-    packageVersion +
+    `${pkg.publisher}.${pkg.name}: ${pkg.version}\n` +
     `Cedar SDK: ${cedar.getCedarSDKVersion()}\n` +
-    vsCodeVersion +
-    nodeVersion;
+    `Visual Studio Code: ${vscode.version}\n` +
+    (process.versions.node ? `node: ${process.versions.node}\n` : '');
 
   const result = await vscode.window.showInformationMessage(
     extensionDetails,